Create simple home page and move dashboard
